refactor: 관리자 기능 수정 및 추가 구현 - 요구사항 조회 시 통합 조회 서비스를 사용하도록 변경 - 반납 기능 구현
